+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
 
+## [Unreleased]
+
+### Added
+- cdk: Add `get_all_mint_info` to MultiMintWallet ([thesimplekid]).
+
 ## [0.14.0](https://github.com/cashubtc/cdk/releases/tag/v0.14.0)
 
 ### Summary